% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_distribution_uis.R
\name{build_dist_tbl}
\alias{build_dist_tbl}
\title{Build Distribution Table}
\usage{
build_dist_tbl(
  distribution,
  n_obs,
  mu = NULL,
  sigma = NULL,
  shape = NULL,
  scale = NULL,
  meanlog = NULL,
  sdlog = NULL
)
}
\arguments{
\item{distribution}{}

\item{n_obs}{}

\item{mu}{}

\item{sigma}{}

\item{shape}{}

\item{scale}{}

\item{meanlog}{}

\item{sdlog}{}
}
\value{

}
\description{
Build Distribution Table
}
