% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_plot.R
\name{plot_dist}
\alias{plot_dist}
\title{Plot Distribution as Histogram}
\usage{
plot_dist(data, x = value, color = distribution, n_obs)
}
\arguments{
\item{data}{(dataframe)}

\item{x}{(var) column from `data` to use as x-value}

\item{color}{(var) column from `color` to specify histogram color}

\item{n_obs}{(numeric) number of observations to use}
}
\value{
ggplot object
}
\description{
Plot Distribution as Histogram
}
