% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_distribution_uis.R
\name{use_gamma_ui}
\alias{use_gamma_ui}
\title{Inputs for Gamma Distribution}
\usage{
use_gamma_ui(session)
}
\arguments{
\item{session}{(session object)}
}
\description{
Inputs for Gamma Distribution
}
