% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_distribution_uis.R
\name{use_lnorm_ui}
\alias{use_lnorm_ui}
\title{Inputs for Log Normal Distribution}
\usage{
use_lnorm_ui(session)
}
\arguments{
\item{session}{}
}
\value{

}
\description{
Inputs for Log Normal Distribution
}
