% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_distribution_uis.R
\name{use_norm_ui}
\alias{use_norm_ui}
\title{Inputs for Normal Distribution}
\usage{
use_norm_ui(session)
}
\arguments{
\item{session}{}
}
\value{

}
\description{
Inputs for Normal Distribution
}
