% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_distribution_uis.R
\name{build_dist_tbl}
\alias{build_dist_tbl}
\title{Build Distribution Table}
\usage{
build_dist_tbl(
  distribution,
  n_obs,
  mu = NULL,
  sigma = NULL,
  shape = NULL,
  scale = NULL,
  meanlog = NULL,
  sdlog = NULL
)
}
\arguments{
\item{distribution}{(string)}

\item{n_obs}{(numeric)}

\item{mu}{(numeric)}

\item{sigma}{(numeric)}

\item{shape}{(numeric)}

\item{scale}{(numeric)}

\item{meanlog}{(numeric)}

\item{sdlog}{(numeric)}
}
\value{

}
\description{
Build Distribution Table
}
